% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{coef.blb4lm}
\alias{coef.blb4lm}
\title{Calculating coefficients from blb4lm}
\usage{
\method{coef}{blb4lm}(object, confidence, level, ...)
}
\arguments{
\item{object}{The output from the function blb4lm}

\item{confidence}{Whether the user needs a confidence interval}

\item{level}{The level of significance}

\item{...}{Additional arguments to be passed to other functions}
}
\value{
The coeffiecients, or the confidence intervals of the coefficients
}
\description{
Calculating coefficients from blb4lm
}
