% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{predict.blb4lm}
\alias{predict.blb4lm}
\title{Predict for new data after doing blb4lm}
\usage{
\method{predict}{blb4lm}(object, new_data, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The output from the function blb4lm}

\item{new_data}{A data frame}

\item{confidence}{Whether the user needs a confidence interval (TRUE/FALSE). Default is 'FALSE'.}

\item{level}{The level of significance. Default is 0.95.}

\item{...}{Additional arguments to be passed to other functions}
}
\value{
sigma, or the confidence intervals of the sigma
}
\description{
Predict for new data after doing blb4lm
}
