% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb4lm_user.R
\name{print.blb4lm}
\alias{print.blb4lm}
\title{Print the formula of blb4lm}
\usage{
\method{print}{blb4lm}(x, ...)
}
\arguments{
\item{x}{The output from the function blb4lm}

\item{...}{Additional arguments to be passed to other functions}
}
\value{
The formula of linear regression used in blb4lm
}
\description{
Print the formula of blb4lm
}
