% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb4lm.R
\docType{package}
\name{blb4lm-package}
\alias{blb4lm-package}
\alias{_PACKAGE}
\title{blb4lm: Implement bag of little bootstrap on Linear regression}
\description{
Implement bag of little bootstrap on linear regression model. 
    It returns coefficients, sigma, and prediction given new data along with their confidence intervals.
}
\details{
The package contains function: blb4lm(), coef.blb4lm(), sigma.blb4lm(), predict.blb4lm().
    These functions can help users quickly apply bag of little bootstrap on linear regression
to get the estimates of coefficients and their standard deviation. Also, users can predict
new data with the new model.
}
\author{
\strong{Maintainer}: Yifeng Jiang \email{yifjiang@ucdavis.edu}

Authors:
\itemize{
  \item Zhihao Chen \email{czhchen@ucdavis.edu}
}

Other contributors:
\itemize{
  \item Shenglun Wu \email{shlwu@ucdavis.edu} [contributor]
  \item Qiushuang Xu \email{qsxu@ucdavis.edu} [contributor]
}

}
