% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb4lm.R
\docType{package}
\name{blb4lm-package}
\alias{blb4lm-package}
\alias{_PACKAGE}
\title{blb4lm: Implement bag of little bootstrap on Linear regression}
\description{
Implement bag of little bootstrap on a given linear regression model. 
    The outputs returns coefficients and sigma of the model, and predicted values given new data along with their confidence intervals.
}
\details{
The package contains function: blb4lm(), coef.blb4lm(), sigma.blb4lm(), predict.blb4lm().
}
\author{
\strong{Maintainer}: Yifeng Jiang \email{yifjiang@ucdavis.edu}

Authors:
\itemize{
  \item Zhihao Chen \email{czhchen@ucdavis.edu}
}

Other contributors:
\itemize{
  \item Shenglun Wu \email{shlwu@ucdavis.edu} [contributor]
  \item Qiushuang Xu \email{qsxu@ucdavis.edu} [contributor]
}

}
