% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb4lm.R
\name{blb4lm}
\alias{blb4lm}
\title{Bag of Little Bootstrap for Linear Regression}
\usage{
blb4lm(formula, s, r, data, parallel = TRUE)
}
\arguments{
\item{formula}{The formula for linear regression}

\item{s}{Number of subsamples}

\item{r}{Times of resampling}

\item{data}{A data frame}

\item{parallel}{Whether to use parallel (TRUR/FALSE). True is set as default and the user
should use "plan()" from furrr package first.}
}
\value{
A list of list of coefficients and sigma
}
\description{
Bag of Little Bootstrap for Linear Regression
}
\details{
Split the training data into s subsamples and resampling r times; then fit lm model for each resamples.
The result is stored as a list of list of two vectors.
}
