% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb4lm.R
\name{blb4lm}
\alias{blb4lm}
\title{Bag of Little Bootstrap for Linear Regression}
\usage{
blb4lm(formula, s, r, data, parallel = FALSE)
}
\arguments{
\item{formula}{The formula for linear regression}

\item{s}{Number of subsamples}

\item{r}{Times of resampling}

\item{data}{A data frame}

\item{parallel}{Whether to use parallel (TRUR/FALSE). If you want use parallel, type "plan(multiprocess, workers = num_cores)" in console first.}
}
\value{
A list of list of coefficients and sigma
}
\description{
Bag of Little Bootstrap for Linear Regression
}
\details{
Split the training data into s parts and resampling r times; then fit lm model for each resamples.
}
