% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{blb4lm}
\alias{blb4lm}
\title{Bag of Little Bootstrap for Linear Regression}
\usage{
blb4lm(formula, s, r, data)
}
\arguments{
\item{formula}{The formula for linear regression}

\item{s}{Number of subsamples}

\item{r}{Times of resampling}

\item{data}{A data frame}
}
\value{
A list of list of coefficients and sigma
}
\description{
Bag of Little Bootstrap for Linear Regression
}
\details{
Bag of Little Bootstrap for Linear Regression
}
