% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blb4lm_user.R
\name{coef.blb4lm}
\alias{coef.blb4lm}
\title{Calculating coefficients from blb4lm}
\usage{
\method{coef}{blb4lm}(object, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{The output from the function blb4lm}

\item{confidence}{Whether the user needs a confidence interval (TRUE/FALSE). Default is 'FALSE'.}

\item{level}{The level of significance. Default is 0.95.}

\item{...}{Additional arguments to be passed to other functions}
}
\value{
The coeffiecients, or the confidence intervals of the coefficients
}
\description{
Calculating coefficients from blb4lm
}
