% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{sigma.blb4lm}
\alias{sigma.blb4lm}
\title{Claculating sigma from blb4lm}
\usage{
\method{sigma}{blb4lm}(object, confidence, level, ...)
}
\arguments{
\item{object}{The output from the function blb4lm}

\item{confidence}{Whether the user needs a confidence interval}

\item{level}{The level of significance}

\item{...}{Additional arguments to be passed to other functions}
}
\value{
sigma, or the confidence intervals of the sigma
}
\description{
Claculating sigma from blb4lm
}
